move lexical cloneDeep helper to deepCopyObjectComplex
